fixed wrapper and auth nav
