+++ src/actions/userActions.js
@@ -44,7 +44,7 @@ export default {
                 dispatch({type: 'LOGIN'});
                 dispatch({type: 'SET_LOGIN_USER', email});
                 dispatch({type: 'STOP_SPINNER'});
-                history.go(-1);
+                history.pushState(null, '/');
 
             } else {
                 console.log('not authenticated');
